3.52 (04/30/09)
Various fixes and protocol change for NATPunchthroughServer. Punchthrough recipient now gets ID_NAT_TARGET_UNRESPONSIVE and ID_NAT_CONNECTION_TO_TARGET_LOST at the same time it is sent to the sender.
Added ConnectionGraph2 and FullyConnectedMesh2 plugins.
RakNetTransport2 plugin added, which works as a plugin instead of creating its own instance of RakPeer. RakNetTransport, which did that, is now depreciated.
Deleted FCMHost plugin (replaced by FullyConnectedMesh2).
ReplicaManager3 fix for not sending serialization to new connections when using identical serialization.
Various fixes for Lobby2 system.
Fix crash in LightweightDatabaseServer if not automatically storing system address.
Added DependentExtensions\IrrlichtDemo
Disallow reconnecting when the remote system thinks you are still connected, if using a different IP but the same RakNetGUID.
Added RakPeerInterface::GetSystemList to return connected SystemAddress and RakNetGUID in a list.
Various fixes for DirectoryServer.php. PHPDirectoryServer::ProcessHTTPRead now returns an enumeration, rather than a string, to indicate results.
RakNetTime is now 64 bit by default. Comment __GET_TIME_64BIT in RakNetDefines.h to change to 32 bit.
UDP proxy system now automatically chooses best server. If more than one server is available, and both source and target relay systems are running RakNet, then source and target will ping all available servers. The servers will be attempted in order of lowest ping sum to highest.
Updated FAQ. Updated manual for NATPunchthrough.

+++ Source/EmailSender.cpp
@@ -148,7 +148,7 @@ const char *EmailSender::Send(const char *hostAddress, unsigned short hostPort, 
 	int i,j;
 	if (attachedFiles && attachedFiles->fileList.Size())
 	{
-		seedMT(RakNet::GetTime());
+		seedMT((unsigned int) RakNet::GetTime());
 		// Random multipart message boundary
 		for (i=0; i < boundarySize; i++)
 			boundary[i]=base64Map[randomMT()%64];
